Do not assume php 7 within the main plugin file.
Most users trying to install the plugin in an host with php5 will not
get the right information we defined in the main plugin file about the
php version, they instead get a `Parser Error` see #85

To avoid problems like that where users can get a white screen I suggest
to not assume php 7 within the main plugin file and use php5 code.

This will ensure users will get the admin notice message and not a white
screen.

More over, in future WordPress will have the WSOD protection but I guess
it's better for the plugin do not rise that protection and instead give
a better information about why the plugin cannot work with their Wp
installation.